requirements.txt
selenium==4.18.1
pytest==8.0.2
pytest-html==4.1.1
pyyaml==6.0.1
webdriver-manager==4.0.1